+++ headers/private/kernel/arch/sparc/arch_platform.h
@@ -33,7 +33,7 @@ public:
 	virtual char SerialDebugGetChar() = 0;
 	virtual void SerialDebugPutChar(char c) = 0;
 
-	virtual	void SetHardwareRTC(uint32 seconds) = 0;
+	virtual	void SetHardwareRTC(uint64 seconds) = 0;
 	virtual	uint32 GetHardwareRTC() = 0;
 
 	virtual	void ShutDown(bool reboot) = 0;
